refactor: extract runtime_builder module and shared write_response helper
Addresses two code review points on the on_gateway_error PR:

1. The spawn+validate pattern was duplicated between build_operation_interceptors
   and resolve_global_error_hook, and the latter didn't belong in path_match/mod.rs.
   Extracted to path_match/runtime_builder.rs following the module_resolver pattern:
   - spawn_interceptor_runtime(): spawns a Node.js process for a resolved module
   - validate_hook(): calls validate(), silently ignoring FunctionNotFound
   - resolve_global_error_hook(): moved here from mod.rs, uses the above helpers

2. The ResponseHeader build/write pattern was duplicated between gateway_error.rs
   and upstream_plugin/mod.rs Step E. Extracted to proxy_utils::write_response(),
   shared by both callers (gateway_error uses .ok(), plugin dispatch uses ?).
