+++ plenum-core/src/path_match/mod.rs
@@ -1,4 +1,5 @@
 mod module_resolver;
+mod runtime_builder;
 
 use std::collections::{BTreeMap, HashMap};
 use std::error::Error;
@@ -182,6 +183,7 @@ fn build_operation_interceptors(
         let resolved = module_resolver::resolve_module(&config.module, config_base)
             .map_err(|e| format!("path '{}': interceptor '{}': {}", path, config.module, e))?;
         let cache_key = resolved.cache_key();
+        let context = format!("path '{}': interceptor '{}'", path, config.module);
 
         // Deduplicate: reuse handle if this module was already spawned.
         let runtime = match runtime_cache.entry(cache_key) {
@@ -206,43 +208,8 @@ fn build_operation_interceptors(
                     .clone()
                     .map(|p| p.into_runtime_permissions())
                     .unwrap_or_default();
-
-                let h: Arc<dyn PluginRuntime> = match &resolved {
-                    module_resolver::ResolvedModule::File(file_path) => {
-                        // File-based interceptors run in a sandboxed Node.js child process.
-                        Arc::new(
-                            plenum_js_runtime::external::spawn_sync(
-                                file_path.to_string_lossy().as_ref(),
-                                serde_json::json!({}),
-                                permissions.clone(),
-                            )
-                            .map_err(|e| {
-                                format!(
-                                    "path '{}': interceptor '{}': failed to spawn Node.js runtime: {}",
-                                    path, config.module, e
-                                )
-                            })?,
-                        )
-                    }
-                    module_resolver::ResolvedModule::Internal {
-                        path: module_path, ..
-                    } => {
-                        // Built-in interceptors run in Node.js via the node-runtime.
-                        Arc::new(
-                            plenum_js_runtime::external::spawn_sync(
-                                module_path.to_string_lossy().as_ref(),
-                                serde_json::json!({}),
-                                permissions.clone(),
-                            )
-                            .map_err(|e| {
-                                format!(
-                                    "path '{}': interceptor '{}': failed to spawn Node.js runtime: {}",
-                                    path, config.module, e
-                                )
-                            })?,
-                        )
-                    }
-                };
+                let h =
+                    runtime_builder::spawn_interceptor_runtime(&resolved, permissions, &context)?;
                 e.insert(h).clone()
             }
         };
@@ -253,17 +220,7 @@ fn build_operation_interceptors(
             .unwrap_or(default_timeout);
 
         let validate_arg = serde_json::to_value(config).unwrap();
-        match runtime.call_blocking("validate", validate_arg, None, timeout) {
-            Ok(_) => {}
-            Err(plenum_js_runtime::JsError::FunctionNotFound(_)) => {}
-            Err(e) => {
-                return Err(format!(
-                    "path '{}': interceptor '{}': validate() failed: {}",
-                    path, config.module, e
-                )
-                .into());
-            }
-        }
+        runtime_builder::validate_hook(runtime.as_ref(), validate_arg, timeout, &context)?;
 
         let hook_handle = HookHandle {
             runtime,
@@ -284,86 +241,6 @@ fn build_operation_interceptors(
     }
 
     Ok(interceptors)
-}
-
-/// Resolve the global `on_gateway_error` interceptor from ServerConfig.
-fn resolve_global_error_hook(
-    server_config: &ServerConfig,
-    config_base: &Path,
-    runtime_cache: &mut HashMap<module_resolver::ModuleCacheKey, Arc<dyn PluginRuntime>>,
-    default_timeout: Duration,
-) -> Result<Option<HookHandle>, Box<dyn Error>> {
-    let cfg = match &server_config.on_gateway_error {
-        Some(c) => c,
-        None => return Ok(None),
-    };
-
-    let resolved = module_resolver::resolve_module(&cfg.module, config_base)
-        .map_err(|e| format!("on_gateway_error interceptor '{}': {}", cfg.module, e))?;
-    let cache_key = resolved.cache_key();
-
-    let runtime = match runtime_cache.entry(cache_key) {
-        std::collections::hash_map::Entry::Occupied(e) => e.get().clone(),
-        std::collections::hash_map::Entry::Vacant(e) => {
-            let permissions = cfg
-                .permissions
-                .clone()
-                .map(|p| p.into_runtime_permissions())
-                .unwrap_or_default();
-
-            let module_path = match &resolved {
-                module_resolver::ResolvedModule::File(p) => p.to_string_lossy().into_owned(),
-                module_resolver::ResolvedModule::Internal { path: p, .. } => {
-                    p.to_string_lossy().into_owned()
-                }
-            };
-
-            let h: Arc<dyn PluginRuntime> = Arc::new(
-                plenum_js_runtime::external::spawn_sync(
-                    &module_path,
-                    serde_json::json!({}),
-                    permissions,
-                )
-                .map_err(|e| {
-                    format!(
-                        "on_gateway_error interceptor '{}': failed to spawn Node.js runtime: {}",
-                        cfg.module, e
-                    )
-                })?,
-            );
-            e.insert(h).clone()
-        }
-    };
-
-    let timeout = cfg
-        .timeout_ms
-        .map(Duration::from_millis)
-        .unwrap_or(default_timeout);
-
-    // Call validate() if exported.
-    let validate_arg = serde_json::json!({
-        "module": &cfg.module,
-        "hook": "on_gateway_error",
-        "function": &cfg.function,
-    });
-    match runtime.call_blocking("validate", validate_arg, None, timeout) {
-        Ok(_) => {}
-        Err(plenum_js_runtime::JsError::FunctionNotFound(_)) => {}
-        Err(e) => {
-            return Err(format!(
-                "on_gateway_error interceptor '{}': validate() failed: {}",
-                cfg.module, e
-            )
-            .into());
-        }
-    }
-
-    Ok(Some(HookHandle {
-        runtime,
-        function: cfg.function.clone(),
-        timeout,
-        options: cfg.options.clone(),
-    }))
 }
 
 pub fn build_router(
@@ -663,7 +540,7 @@ pub fn build_router(
             .map_err(|e| format!("path '{}': {}", path, e))?;
     }
     // Resolve the global on_gateway_error interceptor if configured.
-    let error_hook = resolve_global_error_hook(
+    let error_hook = runtime_builder::resolve_global_error_hook(
         &server_config,
         config_base,
         &mut interceptor_runtime_cache,
